--- src/visualizer/sortingvisualizer/component/radixsortboxes/styles.css
@@ -1,52 +1,0 @@
-.data-arr {
-  align-self: center;
-  display: grid;
-  gap: 10px;
-  justify-content: center;
-}
-
-.stack-arr {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
-
-.box {
-  align-items: center;
-  border: 1px solid black;
-  display: flex;
-  font-size: 16px;
-  justify-content: center;
-  /*margin-bottom: 5px;*/
-  padding: 5px;
-  position: relative;
-  width: 45px;
-}
-
-.stack {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  margin-left: 10px;
-}
-
-.stack-boxes {
-  display: flex;
-  flex-direction: column-reverse;
-}
-
-.number {
-  align-items: center;
-  display: flex;
-  flex-direction: row-reverse;
-  justify-content: center;
-}
-
-.number-with-line {
-  border-top: 1px solid black;
-  font-family: 'Open Sans Semibold', sans-serif;
-  font-size: 15px;
-  text-align: center;
-  width: 60px;
-}
